Create an AnyListUser model if it doesn already exist
If model with the current users' username already exists, use the
existing model instead.

The model is extending the basic Model class instead of the User
class since it is giving out an error when saving a model extending
the User class.

Might need to update later to properly use the User class.