Able to delete a list
When the user delete a list, all posts will or should be deleted.
Not sure if this commit does that in all situations even when
there are a lot of posts in the list.

Also, there started changing the routing, getting ready to use
the signingKeyId to determine ownership instead of url routes.